+++ docs/source/tutorials/code/ddpg_from_pixels.py
@@ -1,4 +1,5 @@
 import numpy as np
+import os
 
 import torch
 import torch.nn as nn
@@ -8,7 +9,7 @@ import torch.nn.functional as F
 from mushroom_rl.algorithms.actor_critic import DDPG
 from mushroom_rl.core import Core
 from mushroom_rl.environments.dm_control_env import DMControl
-from mushroom_rl.environments.realistic_envs.igibson_env import iGibson
+from mushroom_rl.environments.igibson_env import iGibson
 from mushroom_rl.policy import OrnsteinUhlenbeckPolicy
 from mushroom_rl.utils.dataset import compute_J
 
@@ -92,7 +93,8 @@ horizon = 500
 gamma = 0.99
 gamma_eval = 1.
 #mdp = DMControl('walker', 'stand', horizon, gamma, use_pixels=True)
-mdp = iGibson(is_discrete=False)
+config_file = os.path.join(igibson.root_path, 'test', 'test_house.yaml')
+mdp = iGibson(config_file=config_file, is_discrete=False)
 
 # Policy
 policy_class = OrnsteinUhlenbeckPolicy
