+++ examples/habitat_dqn/habitat_dqn.py
@@ -18,7 +18,7 @@ from mushroom_rl.utils.dataset import compute_metrics
 from mushroom_rl.utils.parameters import LinearParameter, Parameter
 from mushroom_rl.utils.replay_memory import PrioritizedReplayMemory
 
-from mushroom_rl.environments.realistic_envs import __file__ as path_robots
+from mushroom_rl.examples.habitat_dqn import __file__ as path_example
 
 
 """
@@ -144,21 +144,18 @@ def experiment():
     parser = argparse.ArgumentParser()
 
     arg_env = parser.add_argument_group('Environment')
-    arg_env.add_argument("--config-path",
+    arg_env.add_argument("--config-file",
                           type=str,
-                          default=os.path.join(mushroom_rl.root_path, 'environments', 'realistic_envs', 'pointnav_nomap.yaml'),
+                          default=os.path.join(path_example, 'pointnav_nomap.yaml'),
                           help='yaml config file.')
-
     arg_env.add_argument("--scene-name",
                           type=str,
                           default='apartment_0',
                           help='Replica scene name.')
-
     arg_env.add_argument("--replica-json",
                           type=str,
-                          default=os.path.join(mushroom_rl.root_path, 'environments', 'realistic_envs', 'replica-start.json.gz'),
+                          default=os.path.join(path_example, 'replica-start.json.gz'),
                           help='.')
-
 
     arg_mem = parser.add_argument_group('Replay Memory')
     arg_mem.add_argument("--initial-replay-size", type=int, default=50000,
@@ -305,8 +302,7 @@ def experiment():
         max_steps = args.max_steps
 
     # MDP
-    print(args.config_path)
-    mdp = HabitatNav(args.scene_name, args.config_path, args.replica_json)
+    mdp = HabitatNav(args.scene_name, args.config_file, args.replica_json)
 
     if args.load_path:
         logger = Logger(DQN.__name__, results_dir=None)
